Various improvments, refactoring

+++ wmelon/core/testing/Benchmark.php
@@ -67,7 +67,7 @@ class Benchmark
       
       unset(self::$benchmarks[$benchmarkName]);
       
-      $difference = bcsub($now, $then);
+      $difference = (int) bcsub($now, $then);
       
       // saving difference in DB if $saveToDatabase = true
